feat: disable testnet option on mobile
